: Bumped to

+++ bin/checkAndPush.ps1
@@ -39,13 +39,37 @@ process {
         $man = Resolve-Path $man
         $folder = Split-Path $man -Parent
         $file = Split-Path $man -Leaf
+        # 打印man folder $file
+        Write-Host "Manifest: $man" -ForegroundColor Green
+        Write-Host "Folder: $folder" -ForegroundColor Green
+        Write-Host "File: $file" -ForegroundColor Green
+
+        if (-not (Test-Path $man)) {
+            Write-Host "Manifest not found: $man" -ForegroundColor Red
+            continue
+        }
+
+        Write-Host 'Checking' -ForegroundColor Green
+
         $noExt = ($file -split '\.')[0]
         $cmd = 'checkver'
+        # 打印noExt $cmd $arg
+        Write-Host "noExt: $noExt" -ForegroundColor Green
+        Write-Host "cmd: $cmd" -ForegroundColor Green
+        Write-Host "arg: $arg" -ForegroundColor Green
 
+        if (-not (Test-Path "$PSScriptRoot\$cmd.ps1")) {
+            Write-Host "Script not found: $PSScriptRoot\$cmd.ps1" -ForegroundColor Red
+        }
         if ($Force) { scoop cache rm $noExt }
         if ($Hashes) { $cmd = 'checkhashes' }
 
+        # 打印Hashes
+        Write-Host "Hashes: $Hashes" -ForegroundColor Green
+        Write-Host "$PSScriptRoot\$cmd.ps1 '$noExt' '$folder' $arg" -ForegroundColor Green
+
         Invoke-Expression -Command "$PSScriptRoot\$cmd.ps1 '$noExt' '$folder' $arg"
+
 
         $updated = @(git status -s)
 
@@ -54,6 +78,9 @@ process {
             # $manifest = Get-Content $man -Raw -Encoding UTF8 | ConvertFrom-Yaml -Ordered
             [psobject] $manifest = Get-Content $man -Raw -Encoding UTF8 | ConvertFrom-Json
             $message = "$noExt`: Bumped to $($manifest.version)"
+            # 打印manifest $message
+            Write-Host "manifest: $manifest" -ForegroundColor Green
+            Write-Host "message: $message" -ForegroundColor Green
 
             if ($Hashes) { $message = "${noExt}: Fixed hashes" }
